hackwave-backend/requirements.txt
django
openai
djangorestframework
django-cors-headers
channels
daphne